+++ index.html
@@ -10,8 +10,8 @@
 	<script src="https://code.jquery.com/jquery-3.3.1.min.js" integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8="
 	 crossorigin="anonymous"></script>
 
-<script src="https://cdnjs.cloudflare.com/ajax/libs/howler/2.0.14/howler.min.js" ></script>
-	 
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/howler/2.0.14/howler.min.js"></script>
+
 </head>
 
 <body>
@@ -42,15 +42,17 @@
 <script>var exports = {};</script>
 <script>var notes = [];</script>
 <script src="src/js/globals.js"></script>
-<script src="src/js/getUserIntervals.js "></script>
+<!-- <script src="src/js/sound.js"></script> -->
+<!-- <script src="src/js/getUserIntervals.js "></script> -->
 <!-- <script src="src/js/getInterval.js "></script> -->
-<script src="src/js/processDOMChord.js"></script>
-<script src="src/js/getNoteChord.js"></script>
+<!-- <script src="src/js/processDOMChord.js"></script> -->
+<!-- <script src="src/js/getNoteChord.js"></script> -->
 <!-- <script src="src/js/getNoteId.js"></script> -->
-<script src="src/js/updateChord.js"></script>
+<!-- <script src="src/js/updateChord.js"></script> -->
 <!-- <script src="src/js/getChord.js"></script> -->
 <script src="src/js/external/howler.min.js"></script>
-<script src="src/js/sound.js"></script>
-<script src="src/js/index.js"></script>
+
+<!-- <script src="src/js/index.js"></script> -->
+<script src="src/js/bundle.js"></script>
 
 </html>
