+++ src/js/getUserIntervals.js
@@ -1,5 +1,5 @@
-// const getInterval = require('./getInterval').getInterval
-// const getNoteId = require('./getNoteId')
+const getInterval = require('./getInterval').getInterval
+const getNoteId = require('./getNoteId')
 
 //return array of intervals found in chord
 function getUserIntervals(userChord) {
